Add changes

+++ microservices/aggregator-microservice/src/main/java/hr/fer/labpro/rassus/aggregatormicroservice/Controllers/AggregateController.java
@@ -55,7 +55,7 @@ public class AggregateController {
             String hum = humidity.getBody() + " %";
 
             if(temperatureUnit.trim().toLowerCase().equals("k")) {
-                Integer tempInt = Integer.parseInt(temp) + 273;
+                Double tempInt = Double.parseDouble(temp) + 273;
                 temp = tempInt.toString() + " K";
             } else {
                 temp += " C";
